Add visibility prop for manual consent view control
Introduces a new prop, `visibility`, to the BiteConsent component that
allows the users to manually control the visibility state of the cookie
consent view or be automatically managed by itself.

The new prop accepts three possible values:
  - `auto`: Automatic visibility management based on the existance of a
    consent cookie (default)
  - `visible`: Explicitly sets the cookie consent view to be visible
  - `hidden`: Explicitly hides the cookie consent view

This change enhances customization options and aligns with common use
cases for managing cookie consent views.